requirements.txt
python-dotenv
pyTelegramBotAPI
gTTS
openai